added menu for selecting what kind of diagram to create
this commit changes the behaviour of the "new diagram" button, from its
previous functionality of "create a new tiddler with type
text/x-plantuml", to show a menu, listing the available Kroki diagram
types that are supported by this tiddlywiki plugin.

It also (ab)uses the tiddlers used for each content type, to use their
content as a template for each new diagram; for example, for graphviz,
it will create a new tiddler with the "digraph g {}" boilerplate already
in place.

Finally, and somewhat minor/cosmetic, the menu shows in the first place
the diagram type for PlantUML, then Pikchr, then GraphViz, then the
other types, mostly because those are the most frequent diagram types
that I use